+++ src/all_logs_vtab.hh
@@ -60,7 +60,6 @@ private:
     logline_value_meta alv_schema_meta;
     logline_value_meta alv_values_meta;
     logline_value_meta alv_src_meta;
-    logline_value_meta alv_thread_meta;
     logline_value_meta alv_stacktrace_meta;
 };