+++ src/hasher.hh
@@ -87,6 +87,18 @@ public:
         return retval;
     }
 
+    uint64_t to_bloom_bits() const
+    {
+        uint64_t h1;
+        uint64_t h2;
+        uint64_t retval = 0;
+
+        this->h_context.Final(&h1, &h2);
+        retval |= 1UL << (h1 % 56);
+        retval |= 1UL << (h2 % 56);
+        return retval;
+    }
+
     void to_string(auto_buffer& buf) const
     {
         auto bits = this->to_array();